# support tensor trasnform arithmetic mode per-channel op w/ORC