---
# [Template] PR Description

Change gsttensor_transform.c to support per-channel arithmetic op w/ORC

Resolves: #123
See also: #456, #789

**Self evaluation:**
1. Build test: [*]Passed [ ]Failed []Skipped
2. Run test: [*]Passed [ ]Failed []Skipped

When I tested performance of arithmetic per-channel op w/ORC, results is 
3:256:256:1 -> 8ms
256:256:3:1 -> 51micro seconds

without ORC
3:256:256:1 -> 10ms