@anyj0527 2023-04-11
`op_s->applying_ch` is initialized with -1, to support operations applied to "all channels". If there is no number given after `@`, this mode is enabled. Please consider this case. Currently there is no testcase for this usecase...

This is an example:
<add -1.5 to all channels and add 2.5 to the 2-th channel>
```bash
gst-launch-1.0 videotestsrc pattern=2 num-buffers=1 ! videoscale ! videoconvert ! \
video/x-raw,format=RGB,width=2,height=2,framerate=0/1 ! tensor_converter ! \
tensor_transform mode=arithmetic option=typecast:float32,per-channel:true@0,add:1.5,add:2.5@2 ! \
filesink location=out.log && od -vf out.log
...
Pipeline is PREROLLED ...
Setting pipeline to PLAYING ...
New clock: GstSystemClock
Got EOS from element "pipeline0".
Execution ended after 0:00:00.000139785
Setting pipeline to PAUSED ...
Setting pipeline to READY ...
Setting pipeline to NULL ...
Freeing pipeline ...
0000000             1.5             1.5               4             1.5
0000020             1.5               4             1.5             1.5
0000040               4             1.5             1.5               4
0000060
```