@byungs2 2023-04-11
Oh I fully understood what I missed.
Thanks for your comment.
By the way, I wonder if after I fixed code for that case, then should I send new PR?
Or can I push new commit on this PR?